Restructure gem files
