+++ lib/shift.rb
@@ -1,0 +1,12 @@
+require "active_support/core_ext/module/delegation"
+require "logger"
+require "sentry-raven"
+require "newrelic_rpm"
+
+require "shift"
+require "shift/circuit_breaker"
+require "shift/circuit_logger"
+require "shift/circuit_monitor"
+
+module Shift
+end
