+++ lib/shift/version.rb
