+++ shift-circuit-breaker.gemspec
@@ -1,10 +1,12 @@
-$LOAD_PATH.unshift File.expand_path('../lib', __FILE__)
+# -*- encoding: utf-8 -*-
+$:.push File.expand_path("../lib", __FILE__)
 
-require "shift/circuit_breaker/version"
+require "shift/version"
 
 Gem::Specification.new do |s|
   s.name        = "shift-circuit-breaker"
   s.version     = Shift::CircuitBreaker::VERSION
+  s.platform    = Gem::Platform::RUBY
   s.summary     = "A generic implementation of the Circuit Breaker pattern in Ruby"
   s.description = <<-DOC
     The library provides a mechanism for detecting, monitoring and controlling slow or 
@@ -16,10 +18,12 @@ Gem::Specification.new do |s|
   s.date        = Time.now.strftime('%Y-%m-%d')
   s.license     = "MIT"
 
-  s.required_ruby_version = '>= 2.0.0'
+  s.required_ruby_version = ">= 2.0.0"
+  s.require_paths = [ "lib" ]
 
-  s.files         = Dir[ "{lib}/**/**/*.{rb}", "LICENSE", "README.md" ]
-  s.test_files    = Dir[ "spec/**/*" ]
+  s.files         = `git ls-files`.split("\n")
+  s.test_files    = `git ls-files -- { test, spec, features }/*`.split("\n")
+  s.executables   = `git ls-files -- bin/*`.split("\n").map{ |f| File.basename(f) }
 
   s.add_dependency "activesupport", ">= 3.0.0"
   s.add_dependency "newrelic_rpm", "~> 4.7", ">= 4.7.1.340"
